% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_tempo.R
\name{def_tempo}
\alias{def_tempo}
\title{Viewing definitions.}
\usage{
def_tempo(table)
}
\description{
Function to view the definitions of the statistical indices
in R Studio's Viewer pane, after the table was downloaded. Table name is
given by the name of the data frame object loaded in the environment by
after downloading the table.
}
\examples{
\dontrun{
def_tempo("table_name.html")
}
}
\author{
Toma, I., Tiru A., Necula, M.
}
