% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_tempo.R
\name{dir_tempo}
\alias{dir_tempo}
\title{Create INS working directory}
\usage{
dir_tempo()
}
\description{
Function to create INS directory and set it as working directory.
}
\examples{
 \dontrun{
def_tempo("table_name.html")
}
}
\author{
Toma, I., Tiru A., Necula, M.
}
