% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_tempo.R
\name{login_tempo}
\alias{login_tempo}
\title{Login to a Tempo Online session}
\usage{
login_tempo(usr = NULL, passwd = NULL)
}
\arguments{
\item{usr}{A character string supplying a valid e-mail adress, used to create a valid Tempo Online account.}

\item{passwd}{A character string supplying a valid password, used to create a valid Tempo Online account.}
}
\value{
Returns an object of class \code{\link{List}} which stores a valid http session for downloading Tempo Online data.
}
\description{
Login to a new session to access Tempo Online. Requires a valid Tempo Online account. If no arguments are provided returns an error message.
}
\examples{
 \dontrun{
login_tempo(usr = "valid_user.name@mail_server.com", passwd = "valid_uncrackable_password")
}
}
\author{
Toma, I., Tiru A., Necula, M.
}
