% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nav_tempo.R
\name{nav_tempo}
\alias{nav_tempo}
\title{Text-based navigation in Tempo Online}
\usage{
nav_tempo()
}
\value{
Returns an object of class List which stores a http
session based on the user's input.
}
\description{
Navigate in Tempo Online by select the numbers coresponding
to the name of the statistical domain.
}
\examples{
\dontrun{
nav_tempo()
}
}
\author{
Toma, I., Tiru A., Necula, M.
}
