% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_tempo.R
\name{search_tempo}
\alias{search_tempo}
\title{Searching Tempo Online by keywords}
\usage{
search_tempo(search_word = c())
}
\arguments{
\item{search_word}{a vector of one or multiple strings provided by the user for querying Tempo Online}
}
\description{
Function providing capabilities to search Tempo Online
by typing a keyword to be looked upon in Tempo Online.
}
\examples{
 \dontrun{
# starts the download for only one table selected by the user after the function call
search_tempo(search_word = c("population"))
# starts the download in bulk for all the tabels which contain the provided keywords
search_tempo(search_word = c("population", "inflation", "education"))
}
}
\author{
Toma, I., Tiru A., Necula, M.
}
